move default results into component
