+++ src/app/landing/mainSearch.tsx
@@ -10,9 +10,8 @@ import {
     FormControl,
     Tooltip,
 } from '@mui/material';
-import AutoComplete from 'common/components/autocomplete';
+import AutoComplete, { defaultHumanResults, defaultMouseResults } from 'common/components/autocomplete';
 import { Search, InfoOutlined } from '@mui/icons-material';
-import { Result } from '@weng-lab/ui-components';
 
 type MainSearchProps = {
     assembly: "GRCh38" | "mm10";
@@ -21,91 +20,7 @@ type MainSearchProps = {
 
 const MainSearch: React.FC<MainSearchProps> = ({ assembly, handleAssemblyChange }) => {
 
-    const defaultResults: Result[] = useMemo(() => {
-        if (assembly === "GRCh38") {
-            return [
-                {
-                    title: "chr19:44,905,754-44,909,393",
-                    domain: {
-                        chromosome: "chr19",
-                        start: 44905754,
-                        end: 44909393,
-                    },
-                    description: "chr19:44,905,754-44,909,393",
-                    type: "Coordinate",
-                },
-                {
-                    title: "SP1",
-                    description:
-                        "Sp1 Transcription Factor\nENSG00000185591.10\nchr12:53380176-53416446",
-                    domain: {
-                        chromosome: "chr12",
-                        start: 53380176,
-                        end: 53416446,
-                    },
-                    type: "Gene",
-                },
-                {
-                    title: "EH38E3314260",
-                    description: "chr19:50417519-50417853",
-                    domain: {
-                        chromosome: "chr19",
-                        start: 50417519,
-                        end: 50417853,
-                    },
-                    type: "cCRE",
-                },
-                {
-                    title: "rs9466027",
-                    description: "chr6:21298226-21298227",
-                    domain: {
-                        chromosome: "chr6",
-                        start: 21298226,
-                        end: 21298227,
-                    },
-                    type: "SNP",
-                },
-                {
-                    title: "Adiponectin Levels",
-                    description: "Dastani Z\n22479202",
-                    id: "Dastani_Z-22479202-Adiponectin_levels",
-                    type: "Study",
-                },
-            ];
-        } else return [
-            {
-                title: "chr7:19,696,109-19,699,188",
-                domain: {
-                    chromosome: "chr7",
-                    start: 19696109,
-                    end: 19699188,
-                },
-                description: "chr7:19,696,109-19,699,188",
-                type: "Coordinate",
-            },
-            {
-                title: "Sp1",
-                description:
-                    "Sp1 Transcription Factor\nENSMUSG00000001280.13\nchr15:102406143-102436404",
-                domain: {
-                    chromosome: "chr15",
-                    start: 102406143,
-                    end: 102436404,
-                },
-                type: "Gene",
-            },
-            {
-                title: "EM10E1179536",
-                description: "chr7:19698911-19699257",
-                domain: {
-                    chromosome: "chr7",
-                    start: 19698911,
-                    end: 19699257,
-                },
-                type: "cCRE",
-            },
-        ];
-    }, [assembly]);
+    const defaultResults = assembly === "GRCh38" ? defaultHumanResults : defaultMouseResults
 
     return (
         <>
@@ -175,7 +90,6 @@ const MainSearch: React.FC<MainSearchProps> = ({ assembly, handleAssemblyChange 
                                 </IconButton>
                             ),
                         }}
-                        defaultResults={defaultResults}
                         assembly={assembly}
                         id="main-search-component"
                         slotProps={{
